requirements.txt
annotated-doc==0.0.4
anyio==4.13.0
av==17.0.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
ctranslate2==4.7.1
faster-whisper==1.2.1
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
numpy==2.4.4
onnxruntime==1.24.4
packaging==26.0
protobuf==7.34.1
PyAutoGUI==0.9.54
pycparser==3.0
PyGetWindow==0.0.9
Pygments==2.20.0
PyMsgBox==2.0.1
pyperclip==1.11.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.18.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.2.2
python3-xlib==0.15
pytweening==1.2.0
PyYAML==6.0.3
rich==14.3.3
scipy==1.17.1
setuptools==82.0.1
shellingham==1.5.4
sounddevice==0.5.5
sympy==1.14.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
